requirements.txt
flask==2.0.1
werkzeug==2.0.2
flask-cors==3.0.10
pillow==9.4.0 